Update filters and fix translation class

+++ app/modules/common/models/Translate/Adapter/NativeArray.php
@@ -10,7 +10,8 @@
 namespace Application\Common\Translate\Adapter;
 
 
-class NativeArray extends \Phalcon\Translate\Adapter\NativeArray {
+class NativeArray extends \Phalcon\Translate\Adapter\NativeArray
+{
     /**
      * Returns the translation related to the given key
      *
@@ -18,18 +19,19 @@ class NativeArray extends \Phalcon\Translate\Adapter\NativeArray {
      * @param array $placeholders
      * @return string
      */
-    public function query($index, $placeholders = null) {
+    public function query($index, $placeholders = null)
+    {
 
         if (!empty($index) and !array_key_exists($index, $this->_translate)) {
 
             $transdir = \Phalcon\DI::getDefault()->getConfig()->dirs->translations;
-            $new = [ $index => $index . '*' ];
+            $new = [$index => $index . '*'];
             $this->_translate = $new + $this->_translate;
 
             $d = dir($transdir);
 
-            while (($file = $d->read()) !== false){
-                if(!preg_match('/^[a-z]{2}\.php$/', $file)) {
+            while (($file = $d->read()) !== false) {
+                if (!preg_match('/^[a-z]{2}\.php$/', $file)) {
                     continue;
                 }
 
@@ -38,9 +40,12 @@ class NativeArray extends \Phalcon\Translate\Adapter\NativeArray {
 
                 require $dict;
 
-                $messages = $new + $messages;
-                $content = sprintf("<?php\n\n// app/config/translations/%s\n\n\$messages = %s;", $file, var_export($messages, true));
-                file_put_contents($dict, $content);
+                // Check if $index exists in given dictionary:
+                if (!array_key_exists($index, $messages)) {
+                    $messages = $new + $messages;
+                    $content = sprintf("<?php\n\n// app/config/translations/%s\n\n\$messages = %s;", $file, var_export($messages, true));
+                    file_put_contents($dict, $content);
+                }
             }
 
             $d->close();
